improve master key part in deploy.sh , not tested

+++ deploy/deploy.sh
@@ -17,7 +17,13 @@ ssh $HOST "ln -sf deploy/environments/$ENV.env active-environment.env"
 # .env
 scp .env $HOST:~
 ssh $HOST "echo TAG_MODULE_HANDBOOK=$TAG >> .env"
-echo RAILS_MASTER_KEY=$(cat secrets/config/credentials/$ENV.key) | ssh $HOST "cat >> .env"
+RAILS_MASTER_KEY=$(cat secrets/config/credentials/$ENV.key)
+if [ $RAILS_MASTER_KEY = "" ]; then
+  echo "RAILS_MASTER_KEY missing"
+  exit 42
+fi
+ssh $HOST "echo echo RAILS_MASTER_KEY=$RAILS_MASTER_KEY >> .env"
+#echo RAILS_MASTER_KEY=$(cat secrets/config/credentials/$ENV.key) | ssh $HOST "cat >> .env"
 
 ssh $HOST "sudo docker compose down"
 ssh $HOST "sudo docker compose up -d"
